ref/M_Tanaka/requirements.txt
matplotlib
numpy
pillow
torch
torchvision
tensorboard
future
tqdm
opencv
